+++ src/main/java/com/tfunsal/TaskManagement/services/impl/CommentServiceImpl.java
@@ -3,9 +3,15 @@ package com.tfunsal.TaskManagement.services.impl;
 import com.tfunsal.TaskManagement.dto.CommentDto;
 import com.tfunsal.TaskManagement.dto.TaskDto;
 import com.tfunsal.TaskManagement.entities.Comment;
+import com.tfunsal.TaskManagement.entities.Project;
 import com.tfunsal.TaskManagement.entities.Task;
+import com.tfunsal.TaskManagement.entities.User;
+import com.tfunsal.TaskManagement.exception.CommentNotFoundException;
 import com.tfunsal.TaskManagement.exception.NoSuchTaskExistsException;
+import com.tfunsal.TaskManagement.exception.ProjectNotFoundException;
+import com.tfunsal.TaskManagement.exception.UserNotFoundException;
 import com.tfunsal.TaskManagement.repository.CommentRepository;
+import com.tfunsal.TaskManagement.repository.ProjectRepository;
 import com.tfunsal.TaskManagement.repository.TaskRepository;
 import com.tfunsal.TaskManagement.repository.UserRepository;
 import com.tfunsal.TaskManagement.services.CommentService;
@@ -13,6 +19,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -27,8 +34,10 @@ public class CommentServiceImpl implements CommentService {
 
     private final UserRepository userRepository;
 
+    private final ProjectRepository projectRepository;
 
-    public List<CommentDto> getAllComments(){
+
+    public List<CommentDto> getAllComments() {
         List<Comment> comments = commentRepository.findAll();
         return comments.stream().map(Comment::getDto).collect(Collectors.toList());
     }
@@ -44,15 +53,102 @@ public class CommentServiceImpl implements CommentService {
         return comments.stream().map(Comment::getDto).collect(Collectors.toList());
     }
 
-    public TaskDto addCommentToTaskByUser(Long userId, Long taskId, CommentDto commentDto) {
+    @Override
+    public List<CommentDto> getCommentsByTaskIdAndProjectId(Long projectId, Long taskId) {
+        List<CommentDto> commentsInProjectAndTask = new ArrayList<>();
 
+        Project project = projectRepository.findById(projectId)
+                .orElseThrow(() -> new ProjectNotFoundException("Project not found with id : " + projectId));
 
-        Optional<Task> optionalTask = taskRepository.findById(taskId);
+        Task task = taskRepository.findById(taskId)
+                .orElseThrow(() -> new NoSuchTaskExistsException("Task not found with id :" + taskId));
 
-        if (optionalTask.isPresent()) {
-            Task existingTask = optionalTask.get();
+        if (!task.getProject().getId().equals(projectId)) {
+            throw new IllegalArgumentException("Task does not belong to the specified project.");
+        }
 
-            if (existingTask.getAssignee() != null && existingTask.getAssignee().getId().equals(userId)) {
+        List<CommentDto> allCommentsForTask = getCommentsByTask(taskId);
+
+        for (CommentDto comment : allCommentsForTask) {
+            if (comment.getTaskId().equals(taskId)) {
+                commentsInProjectAndTask.add(comment);
+            }
+        }
+
+        return commentsInProjectAndTask;
+    }
+
+    @Override
+    public List<CommentDto> getUserCommentsByTaskIdAndProjectId(Long projectId, Long taskId, Long userId) {
+        List<CommentDto> userCommentsInProjectAndTask = new ArrayList<>();
+
+        Project project = projectRepository.findById(projectId)
+                .orElseThrow(() -> new ProjectNotFoundException("Project not found with id : " + projectId));
+
+        Task task = taskRepository.findById(taskId)
+                .orElseThrow(() -> new NoSuchTaskExistsException("Task not found with id :" + taskId));
+
+        User user = userRepository.findById(userId).
+                orElseThrow(() -> new UserNotFoundException("User not found with id : " + userId));
+        if (!task.getProject().getId().equals(projectId)) {
+            throw new IllegalArgumentException("The task is not associated with the specified project.");
+        }
+
+        if (!task.getAssignee().getId().equals(userId)) {
+            throw new IllegalArgumentException("The task is not assigned to the specified user.");
+        }
+
+        List<CommentDto> allCommentsForTask = getCommentsByTask(taskId);
+
+        for (CommentDto comment : allCommentsForTask) {
+            if (comment.getTaskId().equals(taskId) && comment.getUser().equals(user.getName())) {
+                userCommentsInProjectAndTask.add(comment);
+            }
+        }
+
+        return userCommentsInProjectAndTask;
+    }
+
+    public TaskDto addCommentToTaskByUser(Long userId, Long projectId, Long taskId, CommentDto commentDto) {
+
+        Optional<Project> project = projectRepository.findById(projectId);
+        if (project.isPresent()) {
+
+            Optional<Task> optionalTask = taskRepository.findById(taskId);
+            if (optionalTask.isPresent()) {
+                Task existingTask = optionalTask.get();
+
+                if (existingTask.getAssignee() != null && existingTask.getAssignee().getId().equals(userId)) {
+
+                    Comment comment = new Comment();
+
+                    comment.setAuthor(userRepository.findById(userId).get());
+                    comment.setContent(commentDto.getContent());
+                    comment.setCreatedDate(LocalDateTime.now());
+                    comment.setTask(existingTask);
+
+                    existingTask.getComments().add(comment);
+
+                    taskRepository.save(existingTask);
+
+                    return existingTask.getDto();
+                }
+            } else {
+                throw new NoSuchTaskExistsException("Task not found with id : " + taskId);
+            }
+        }
+        throw new ProjectNotFoundException("Project not found with id:" + projectId);
+
+    }
+
+    public TaskDto addCommentToTaskByAdmin(Long userId, Long projectId, Long taskId, CommentDto commentDto) {
+
+        Optional<Project> project = projectRepository.findById(projectId);
+        if (project.isPresent()) {
+
+            Optional<Task> optionalTask = taskRepository.findById(taskId);
+            if (optionalTask.isPresent()) {
+                Task existingTask = optionalTask.get();
 
                 Comment comment = new Comment();
 
@@ -66,66 +162,65 @@ public class CommentServiceImpl implements CommentService {
                 taskRepository.save(existingTask);
 
                 return existingTask.getDto();
+
+            } else {
+                throw new NoSuchTaskExistsException("Task not found with id : " + taskId);
+            }
+        }
+        throw new ProjectNotFoundException("Project not found with id: " + projectId);
+    }
+
+
+    public CommentDto updateComment(Long projectId, Long taskId, Long commentId, CommentDto commentDto) {
+        Optional<Project> project = projectRepository.findById(projectId);
+
+        if (project.isPresent()) {
+
+            Optional<Task> optionalTask = taskRepository.findById(taskId);
+
+            if (optionalTask.isPresent()) {
+
+                Task task = optionalTask.get();
+                Optional<Comment> optionalComment = commentRepository.findById(commentId);
+
+                if (optionalComment.isPresent()) {
+                    Comment existingComment = optionalComment.get();
+
+                    existingComment.setContent(commentDto.getContent());
+                    existingComment.setTask(task);
+
+                    return commentRepository.save(existingComment).getDto();
+                } else {
+                    throw new CommentNotFoundException("Comment not found with id: " + commentId);
+                }
+            } else {
+                throw new NoSuchTaskExistsException("Task not found with id: " + taskId);
             }
         } else {
-            throw new NoSuchTaskExistsException("Task not found with id : " + taskId);
+            throw new ProjectNotFoundException("Project not found with id: " + projectId);
         }
-        return null;
     }
 
-    public TaskDto addCommentToTaskByAdmin(Long userId , Long taskId, CommentDto commentDto) {
+    @Override
+    public boolean deleteComment(Long projectId, Long taskId, Long commentId) {
 
-        Optional<Task> optionalTask = taskRepository.findById(taskId);
+        Optional<Project> optionalProject = projectRepository.findById(projectId);
 
-        if (optionalTask.isPresent()) {
-            Task existingTask = optionalTask.get();
-
-            Comment comment = new Comment();
-
-            comment.setAuthor(userRepository.findById(userId).get());
-            comment.setContent(commentDto.getContent());
-            comment.setCreatedDate(LocalDateTime.now());
-            comment.setTask(existingTask);
-
-            existingTask.getComments().add(comment);
-
-            taskRepository.save(existingTask);
-
-            return existingTask.getDto();
-
+        if (optionalProject.isPresent()) {
+            Optional<Task> optionalTask = taskRepository.findById(taskId);
+            if (optionalTask.isPresent()) {
+                Optional<Comment> optionalComment = commentRepository.findById(commentId);
+                if (optionalComment.isPresent()) {
+                    commentRepository.deleteById(commentId);
+                    return true;
+                } else {
+                    throw new CommentNotFoundException("Comment not found with id: " + commentId);
+                }
+            } else {
+                throw new NoSuchTaskExistsException("Task not found with id: " + taskId);
+            }
         } else {
-            throw new NoSuchTaskExistsException("Task not found with id : " + taskId);
+            throw new ProjectNotFoundException("Project not found with id: " + projectId);
         }
-    }
-
-
-    public CommentDto updateComment(Long taskId , Long commentId, CommentDto commentDto) {
-
-        Comment existingComment = commentRepository.findById(commentId).get();
-
-        existingComment.setContent(commentDto.getContent());
-        existingComment.setCreatedDate(LocalDateTime.now());
-
-        Optional<Task> optionalTask = taskRepository.findById(taskId);
-
-        if (optionalTask.isPresent()) {
-            Task task = optionalTask.get();
-            existingComment.setTask(task);
-
-        } else {
-            throw new NoSuchTaskExistsException("Task not found with id : " + commentDto.getTaskId());
-        }
-        return commentRepository.save(existingComment).getDto();
-    }
-
-    public boolean delete(Long commentId) {
-
-        Optional<Comment> optionalComment = commentRepository.findById(commentId);
-
-        if (optionalComment.isPresent()) {
-            commentRepository.deleteById(commentId);
-            return true;
-        }
-        return false;
     }
 }
